requirements.txt
flask_wtf
flask_session
readability-lxml
openai
python-dotenv
flask
beautifulsoup4
requests